+++ lib/ex_unit/lib/ex_unit/callbacks.ex
@@ -127,7 +127,7 @@ defmodule ExUnit.Callbacks do
   ## Helpers
 
   def __merge__(_mod, other, :ok), do: other
-  def __merge__(_mod, other, { :ok, data }), do: Keyword.merge(other, data)
+  def __merge__(_mod, other, { :ok, data }) when is_list(data), do: Keyword.merge(other, data)
   def __merge__(mod, _, failure) do
     raise "expected ExUnit callback in #{inspect mod} to return :ok " <>
           " or { :ok, data }, got #{inspect failure} instead"